parsing MIME message files
using ragel and an adhoc parser, we parse the From, To, Sender,
In-Reply-To, Message-Id headers.

We do not use RFC5322 parsers because after testing on a 110k mail
corpus from 1999 on, it turns out a lot of emails are not standard
compliant.

The parser is packaged inside a dynamic library:

After each parse, we check whether to reload the library and reload it
in case the parser has been updated.