Implement file loading API, async and synch on OSX
On OSX, parsing is done on all cores using dispatch_apply, while I/O is
performed in the background.

We implement parallelism of the main program.

The limitations are that processing tasks are now more memory
constrained, which means that our macroman workaround tasks might not
always have enough memory.

TODO: we could adapt to total memory. Devise parallelism so that it
gives enough memory for each.

* File Loading API (zwei_files.hpp)

We attempted to use the dispatch_io API however it loses to our simple
background thread loading all the data. Dispatch I/O is not particularly
smart about doing reads serially from a rotary disk for instance.

Add test program computing the sha1 of files, which we can easily
compare for performance and reliability with a reference.